refactor: add action types, refactor
