+++ src/constants/context.ts
@@ -3,6 +3,7 @@ export enum Context {
   PLAYER = 'player',
   LIBRARY = 'library',
   ANALYTICS = 'analytics',
+  UNKNOWN = 'unknown',
 
   /** @deprecated */
   EXPLORER = 'explorer',
