+++ src/services/actions/actionTypes.ts
@@ -1,0 +1,11 @@
+export enum ActionType {
+  COPY_ITEM = 'copy_item',
+  CREATE_ITEM = 'create_item',
+  DELETE_ITEM = 'delete_item',
+  MOVE_ITEM = 'move_item',
+  UPDATE_ITEM = 'update_item',
+
+  CREATE_ITEM_MEMBERSHIP = 'create_item_membership',
+  DELETE_ITEM_MEMBERSHIP = 'delete_item_membership',
+  UPDATE_ITEM_MEMBERSHIP = 'update_item_membership',
+}
